Fix same messages being fetched with CHATHISTORY
The issue was that Message.Time() converted the timestamp to Local time,
and this local timestamp was being used with UTC timestamps.  Thus
senpai now converts the time only during display.

Moreover, to avoid missing messages in history (and at the cost of
duplicates), the condition in bufferList.AddLines() as been modified.