Eager registration
Decreases the number of roundtrips necessary for registration.

A typical registration might look like this:

- C: CAP LS; NICK; USER
- S: CAP LS
- C: CAP REQ
- S: CAP ACK
- C: AUTHENTICATE <mesh>
- S: AUTHENTICATE +
- C: AUTHENTICATE <user/pass payload>
- S: RPL_SASLSUCCESS
- C: CAP END
- S: welcome batch

By sending client messages eagerly, despite the number of message not
decreasing (it's increasing when the client supports more caps than the
server), the number of roundtrips is decreased to 1 (not counting TCP
and TLS negotiation).