Support for soju.im/bouncer-networks
This patch also disable the highlight on reconnect. This might be an
issue (the user would want to know when senpai is online again?), but
with multiple connections, it's bothersome to have to unread all of
them on start (it wasn't a problem with only one connection since it was
read instantly).

Now, lastbuffer.txt also contains the network ID, otherwise the user
might end up on another buffer with the same name.

This patch does not extend /r to support multiple networks (it will send
the message to the latest query, whatever the current displayed network
is).