Fix requesting CHATHISTORY of empty channels
Channels which have never received any message return no messages on the
initial chathistory, thus returning an empty bound with first == last ==
time.Time{}.

However this empty bound was stored in the channel bounds, and was used
as starting point for fetching history if Page Up was pressed. This in
turn caused senpai to send a CHATHISTORY BEFORE 01-01-0001 message,
which is invalid.

This fixes the issue by not storing an empty bound if the CHATHISTORY
request does not return any message.